Replace _TaskDecorators class with module level function
